+++ js/util-generating.js
@@ -1,0 +1,9 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.lenght - 1)];
+
+export {getRandomInteger,getRandomArrayElement};
